server/requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient
djangorestframework-simplejwt
requests
python-dotenv
